Added wind. Fix bug where banana would explode when hitting text.

+++ aux.c
@@ -3,7 +3,7 @@
 
 #include <graphics.h>
 
-float myrand(int min, int max)
+float myrand(float min, float max)
 {
     return min + (((rand()+0.0)/(RAND_MAX+0.0))*(max-min));
 }
